Create urdb_client.py
This method provides API integration with the URDB to be able to pull down specific utilities rate structures and within certain date ranges. It runs and if it hits the api limit it backsoff retries and caches the outcome in case of failure. note this does take a while if you are trying to do county level granularity