Create geo_mapping.py
Creates a class for geo mapping capabilities for the utilities in the URBD. This includes utilizing the URBD API to find utilities in different locations. Then you can create the county to utility mapping and the ability to fill in a missing utility with the nearest and vice versa finding missing counties from your data.