Create pricing.py
Creates a class that helps calculate effective rates from the URDB utilizing a demand profile from a specific building. This also includes the ability to calculate the energy and demand and flat rate charges individually. Essentailly this class allows you to take data from a real building and get a sense of the cost of different rates applied to that building profile with some assumptions baked in