Create enhanced_workflow.py
This example is an integration of my existing workflow that inspired this package. This adds in some additional bracketing to rate collection, including energy, demand, and rate type filtering. this also adds in some  natural gas comparisons and specific rate comparisons to determine the "Spark Gap" between NG and electric prices.